add support C++ compilation, moved av_request, av_reply to other file
