+++ src/libav/av.h
@@ -33,15 +33,25 @@ struct av_event {
     int cache;
 };
 
+#ifdef __cplusplus
+extern "C" {
+#endif
+
 int av_register(struct av_connection *conn);
 int av_unregister(struct av_connection *conn);
 int av_register_trusted(struct av_connection *conn);
 int av_unregister_trusted(struct av_connection *conn);
+ssize_t av_parse_request_from_buf(struct av_event *event, const char* buf, size_t size);
 int av_request(struct av_connection *conn, struct av_event *event, int timeout);
+ssize_t av_set_reply_to_buf(char* buf, size_t size, const struct av_event *event);
 int av_reply(struct av_connection *conn, struct av_event *event);
 int av_set_result(struct av_event *event, int res);
 int av_set_cache(struct av_event *event, int cache);
 int av_get_filename(struct av_event *event, char *buf, int size);
+
+#ifdef __cplusplus
+}
+#endif
 
 #endif
